requirements_linux.txt
# Aurum — Dependencias para Linux (GCP VM)
# Generado para Python 3.11+
# NOTA: MetaTrader5 NO está incluido (no compatible con Linux).
#       El shim MetaTrader5/__init__.py usa metaapi-cloud-sdk en su lugar.

# Core trading & datos
pandas>=2.0
numpy>=1.24
ta>=0.11

# Base de datos
psycopg2-binary>=2.9

# MetaAPI Cloud (reemplaza MetaTrader5)
metaapi-cloud-sdk>=25.0

# Inteligencia Artificial
google-genai>=1.0

# Noticias / RSS
feedparser>=6.0
requests>=2.28

# Telegram
python-telegram-bot>=20.0

# Utilidades
python-dotenv>=1.0
psutil>=5.9
pytz>=2023.3
python-dateutil>=2.8
aiohttp>=3.8
tenacity>=8.0
rich>=13.0
APScheduler>=3.10

# Análisis técnico / matemático
scipy>=1.10
